Bear Finished, Waterfall redo, Scoretest bug fix
Bear now walks to, picks up, and consumes fish. correctly turning them
off. if it is a player or not. disables player and renables on reseting
check pt. Layers and placement all fixed
Waterfall was using Mike's Beta version animations, was not working for
me.  Brought it back to this version
Score screen has check to prevent bug
Player returns to checkpt is broken