lessons slide 3, frog cinematic
inserted line at beginning of each lesson slide 3, restating the role
you'll be playing in this level, before the text goes on to set the
scene in concrete second-person terms.

changed frog cinematic so that the puddle doesn't shrink and expand, but
instead just fades in and out. the variables for the timing of these
transitions can be tweaked.